research_lab/requirements.txt
# Core LangChain/LangGraph
langchain>=0.2.0
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# Vector Database
chromadb>=0.5.0

# Data Validation
pydantic>=2.0
pydantic-settings>=2.0

# UI
streamlit>=1.35.0

# Research Tools
arxiv>=2.1.0
semanticscholar>=0.8.0
biopython>=1.83

# HTTP & Async
httpx>=0.27.0
aiohttp>=3.9.0

# Environment & Config
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.0
tiktoken>=0.7.0

